+++ x/posts/internal/types/poll_answers_response_test.go
@@ -11,7 +11,7 @@ import (
 func TestPollAnswersQueryResponse_String(t *testing.T) {
 	testOwner, _ := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
 	answers := types.AnswersDetails{
-		Answers: []uint64{1, 2},
+		Answers: []uint{1, 2},
 		User:    testOwner,
 	}
